SpreadsheetCSV/pyproject.toml
[project]
name = "spreadsheetcsv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=1.7.0",
    "ipykernel>=7.1.0",
    "python-dotenv>=1.1.1",
]
